Move chop to its own file
