+++ flatgfa/src/ops/mod.rs
@@ -1,3 +1,4 @@
+pub mod chop;
 pub mod depth;
 pub mod extract;
 pub mod position;
